Add CSS checkbox demo (#516)
