+++ components/demos/index.jsx
@@ -3,7 +3,7 @@ export { default as AccordionDemo } from './Accordion/stitches';
 export { default as AlertDialogDemo } from './AlertDialog/stitches';
 export { default as AspectRatioDemo } from './AspectRatio/stitches';
 export { default as AvatarDemo } from './Avatar/stitches';
-export { default as CheckboxDemo } from './Checkbox';
+export { default as CheckboxDemo } from './Checkbox/stitches';
 export { default as CollapsibleDemo } from './Collapsible';
 export { default as ContextMenuDemo } from './ContextMenu';
 export { default as DialogDemo } from './Dialog';
