+++ data/primitives/components/checkbox/1.0.1.mdx
@@ -15,9 +15,7 @@ aria: https://www.w3.org/WAI/ARIA/apg/patterns/checkbox
   <CheckboxDemo />
 </HeroContainer>
 
-```jsx hero template=Checkbox
-
-```
+<HeroCodeBlock folder="Checkbox" />
 
 <Highlights
   features={[
